Split net.imglib2.meta to new imglib2-meta module
The rationale is that we are planning to expand the metadata class
hierarchies, adding new Axis subclasses and new interfaces. Some of this
work will introduce at least one new dependency on a units management
system (probably UOMo). To avoid inflicting that dependency on projects
that do not care about such metadata, we separate things out.