Downgrade imglib2-io to use Bio-Formats 4.4 API
The 4.4.x stable releases are what OME is officially supporting, and are
less volatile than the 4.5.x development code (which has still not seen
a release).

This essentially reverts the following two commits:
  * 86fbb6706408e8788f1e08c43ea9c16dc50b99d3
  * 6b9c07e80585b293084c70330fef17fc458c300d

Note that this change is rather temporary, and mainly for the impending
2.0.0-beta-7 release. After that, we will switch to using the new SCIFIO
API (see https://github.com/scifio/scifio for details). The only reason
we don't switch right at this moment is that the new SCIFIO is still
missing one key feature: it cannot lean on Bio-Formats 4.4.x to read
proprietary file formats which have not yet converted to the new API.

Once SCIFIO supports such a fusion, the imglib2-io project will go away
in favor of ImgOpener and ImgOpener support directly in scifio.