Remove PixLib from the performance benchmarks
The inclusion of PixLib nearly doubles the number of lines in the
graphs, making the data much harder to read and interpret.

PixLib is not included with ImageJ2 nor Fiji, nor is it (to my
knowledge) used by any third party ImageJ update sites. So it is
not really a relevant comparison point. If that changes, we can
revert this commit later.